+++ packages/components/pie-notification/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @justeattakeaway/pie-notification
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [[`4f89146dc`](https://github.com/justeattakeaway/pie/commit/4f89146dc83c7eb72cecc0b83be6d718aa395ae1)]:
+  - @justeattakeaway/pie-webc-core@0.19.0
+
 ## 0.3.2
 
 ### Patch Changes